Move flow-based importers to a module and simplify
Extract from_eraa/from_jao/from_tso (and their helpers) into a dedicated
FlowBasedImportersMixin in components/_types/_flow_based_io.py, leaving
flow_based_constraints.py to hold only the component. Trim the module,
class, importer and calculate_zonal_PTDF docstrings to essentials now that
the math lives in the user guide.

Simplifications:
- io.py: extract _export_zonal_ptdf / _import_zonal_ptdf helpers, removing
  the inline prefix blocks and folding the two import paths into one call.
- flow_based.py: vectorise _corridor_cut (get_dummies instead of a per-link
  loop, dropping a triple _classify_columns recompute) and _classify_columns
  (isin masks instead of set-membership comprehensions).
- Replace sphinx :func: roles with plain backticks for mkdocs.

Surface the inherited importers on the API page via inherited_members.
