+++ docs/api/components/types/flow-based-constraints.md
@@ -5,3 +5,8 @@ SPDX-License-Identifier: CC-BY-4.0
 -->
 
 ::: pypsa.components.FlowBasedConstraints
+    options:
+      inherited_members:
+        - from_eraa
+        - from_jao
+        - from_tso
