+++ pypsa/optimization/flow_based.py
@@ -4,29 +4,15 @@
 
 """Flow-based market-coupling constraints.
 
-A flow-based domain constrains the *net positions* of market zones (buses) by a set of
-linear inequalities ``zonal_ptdf . NP <= RAM``. Each row is a critical network element
-(CNEC). The zonal PTDF sensitivities are stored in the dedicated
-``n.c.flow_based_constraints.zonal_ptdf`` frame (cnec x zone, or ``(snapshot, cnec) x zone``
-when time-varying); the ``ram`` attribute (static or time-varying) is the right-hand
-side. Both the PTDF and the RAM may vary by snapshot; the constraint broadcasts either.
+Bounds the net positions of market zones (buses) by ``zonal_ptdf . NP <= RAM``, one row
+per critical network element (CNEC), plus a zero-sum balance ``sum(NP) = 0``. Each net
+position is a variable inside the nodal balance, so zonal prices stay native. A domain
+column may instead name a ``Link`` (an AHC or EvFB HVDC corridor): its flow loads the
+CNECs through its own column while its Core-side balance term is cancelled, so net
+positions stay ``generation - load``.
 
-The net position of a zone is added as a variable directly inside the nodal balance
-(``generation - load - net_position = 0``), so no auxiliary buses or links are needed and
-the zonal prices remain the duals of the nodal balance. A single zero-sum balance
-``sum(NP) = 0`` closes the copper plate across zones (AHC virtual hubs join it, below).
-
-A domain column may also name a ``Link`` instead of a zone bus: the controllable HVDC
-corridors of advanced hybrid coupling (AHC, a border to an external hub) and evolved
-flow-based coupling (EvFB, between two zones). Each such column loads the CNECs through a
-``zonal_ptdf . Link-p`` term, using the existing interconnector (its capacity is the link's
-``p_nom``). To keep every zone's net position equal to ``generation - load``, the corridor
-link's Core-side contribution is cancelled in the nodal balance (EvFB at both ends, AHC at
-its Core end); an AHC border additionally routes its flow onto the ``sum(NP) = 0`` balance,
-where it is the net position of the external virtual hub.
-
-This module is deliberately abstract: it implements linear constraints on grouped bus net
-positions and controllable link flows, and is not specific to any region or framework.
+See the user guide on flow-based market coupling for the formulation and the AHC/EvFB
+treatment.
 """
 
 from __future__ import annotations
@@ -69,59 +55,48 @@ def _classify_columns(n: Network) -> tuple[list, list]:
 
     """
     cols = n.c.flow_based_constraints.zonal_ptdf.columns
-    buses = set(n.c.buses.static.index)
-    links = set(n.c.links.static.index)
-    zone_cols = [c for c in cols if c in buses]
-    link_cols = [c for c in cols if c not in buses and c in links]
-    unknown = [c for c in cols if c not in buses and c not in links]
-    if unknown:
+    is_bus = cols.isin(n.c.buses.static.index)
+    is_link = cols.isin(n.c.links.static.index) & ~is_bus
+    if unknown := cols[~is_bus & ~is_link].tolist():
         msg = (
             f"Flow-based domain columns must be buses (zones) or links, but "
-            f"{unknown} are neither. Add them as buses/links or pass a `buses` mapping."
+            f"{unknown} are neither. Add them as buses/links."
         )
         raise ValueError(msg)
-    return zone_cols, link_cols
+    return cols[is_bus].tolist(), cols[is_link].tolist()
 
 
 def _corridor_cut(n: Network) -> Any:
-    """Per-zone linear expression cancelling each corridor link's Core-side balance term.
+    """Per-zone expression cancelling each corridor link's Core-side balance term.
 
-    The nodal balance adds ``-Link-p`` at ``bus0`` and ``+efficiency . Link-p`` at ``bus1``.
-    Cancelling those terms at a corridor's zone end(s) keeps every zone net position equal to
-    ``generation - load``, so the corridor loads the CNECs only through its own column. A
-    fully external link (neither end a zone) gets a zero column and drops out. Returns a
-    ``(snapshot, name)`` expression over zone buses, or ``None`` if no corridor has a zone
-    end. Its negated bus-sum is the AHC virtual hubs' net exchange with Core (see the plate
-    balance in :func:`define_flow_based_constraints`); internal lossless EvFB nets to zero.
+    The nodal balance adds ``-Link-p`` at ``bus0`` and ``+efficiency . Link-p`` at ``bus1``;
+    cancelling those at a corridor's zone end(s) keeps every zone net position equal to
+    ``generation - load``. Returns a ``(snapshot, name)`` expression over zone buses, or
+    ``None`` if no corridor has a zone end. Its negated bus-sum is the AHC hubs' net
+    exchange with Core (the plate balance in `define_flow_based_constraints`); lossless
+    EvFB nets to zero.
     """
-    zones = set(_classify_columns(n)[0])
-    links = n.c.links.static
-    eff = links["efficiency"]
-    coeff = pd.DataFrame(
-        0.0, index=pd.Index(sorted(zones), name="name"), columns=_classify_columns(n)[1]
-    )
-    for link in coeff.columns:
-        b0, b1 = links.at[link, "bus0"], links.at[link, "bus1"]
-        if b0 in zones:
-            coeff.at[b0, link] += 1.0
-        if b1 in zones:
-            coeff.at[b1, link] += -eff[link]
+    zone_cols, link_cols = _classify_columns(n)
+    if not link_cols:
+        return None
+    links = n.c.links.static.loc[link_cols]
+    zones = pd.Index(sorted(zone_cols), name="name")
+    plus = pd.get_dummies(links["bus0"]).reindex(columns=zones, fill_value=0)
+    minus = pd.get_dummies(links["bus1"]).reindex(columns=zones, fill_value=0)
+    coeff = (plus - minus.mul(links["efficiency"], axis=0)).T
     coeff = coeff.loc[:, (coeff != 0).any()]
     if coeff.columns.empty:
         return None
-    da = xr.DataArray(
-        coeff.values, dims=["name", "link"],
-        coords={"name": coeff.index, "link": coeff.columns},
-    )
+    coeff.columns.name = "link"
     link_p = n.model["Link-p"].sel(name=list(coeff.columns)).rename(name="link")
-    return (link_p * da).sum("link")
+    return (link_p * xr.DataArray(coeff)).sum("link")
 
 
 def flow_based_balance_terms(n: Network, buses: pd.Index) -> Any:
     """Terms the flow-based domain contributes to the nodal balance, or ``None``.
 
     Each zone bus gets ``-net_position`` (its balance reads ``generation - load - NP = 0``)
-    plus the cancellation of any corridor link's Core-side term (see :func:`_corridor_cut`),
+    plus the cancellation of any corridor link's Core-side term (see `_corridor_cut`),
     so net positions stay ``generation - load`` and corridors are not double-counted.
     """
     if NP_VAR not in n.model.variables:
@@ -143,14 +118,11 @@ def _zonal_ptdf(n: Network) -> xr.DataArray:
 
 
 def validate_flow_based(n: Network) -> None:
-    """Check the network is compatible with a flow-based domain.
+    """Reject branches directly connecting two zone buses.
 
-    The domain replaces the electrical grid *between* the market zones, so no branch may
-    directly connect two distinct zone buses, with one exception: a controllable ``Link``
-    that is a declared domain column (an AHC/EvFB HVDC corridor) is kept, since its flow
-    enters the constraint explicitly. Concretely this rejects any cross-zone ``Line`` or
-    ``Transformer`` (passive AC the zonal PTDF already represents) and any cross-zone
-    ``Link`` that is not a domain column.
+    The domain replaces the grid *between* zones, so a cross-zone ``Line``, ``Transformer``
+    or ``Link`` is forbidden, except a ``Link`` that is a declared domain column (an AHC/EvFB
+    corridor), whose flow enters the constraint explicitly.
 
     Raises
     ------
@@ -194,16 +166,11 @@ def define_flow_based_variables(n: Network, sns: pd.Index) -> None:
 
 
 def define_flow_based_constraints(n: Network, sns: pd.Index) -> None:
-    """Define the flow-based constraints ``zonal_ptdf . NP <= RAM`` and ``sum(NP) = 0``.
+    """Define the domain half-spaces ``zonal_ptdf . NP <= RAM`` and the balance ``sum(NP) = 0``.
 
-    The net-position variables are injected into the nodal balance elsewhere; here we add
-    the domain half-spaces and the global balance. Link columns (AHC/EvFB corridors)
-    contribute ``zonal_ptdf . Link-p`` terms, using the link flow in its ``bus0 -> bus1``
-    direction (the domain column must follow that sign convention). The global balance is
-    ``sum(NP) + sum(AHC hub net positions) = 0``: Core zones plus the AHC virtual hubs sum
-    to zero, so Core need not be internally balanced when it exchanges over AHC borders. The
-    domain constraint is indexed by the CNEC (component ``name``), so its dual is mapped to
-    ``mu_domain`` by the standard dual assignment.
+    Link columns (AHC/EvFB corridors) add ``zonal_ptdf . Link-p`` terms in the link's
+    ``bus0 -> bus1`` direction. AHC hubs join the balance via the corridor cut; the domain
+    dual is written to ``mu_domain``.
     """
     if not _has_flow_based(n):
         return
